requirements.txt
streamlit
pytube
moviepy
openai-whisper
torch
yt-dlp
